+++ engine/src/calmar/platform/glfw/window_glfw.hpp
@@ -33,7 +33,7 @@ namespace calmar {
 
         virtual bool closeRequested() const override;
 
-        virtual double getDeltaTime() const override;
+        virtual float getDeltaTime() const override;
 
         virtual void startTiming() override;
         
@@ -46,9 +46,9 @@ namespace calmar {
        private:
         GLFWwindow* mBackendHandle;
 
-        double mDeltaTime = 0.0;
-        double mLastFrameTime = 0.0f;
-        double mCurrentFrameTime = 0.0f;
+        float mDeltaTime = 0.0;
+        float mLastFrameTime = 0.0f;
+        float mCurrentFrameTime = 0.0f;
         u32 mFrameCounter = 0;
         float mFps = 0;
